EPhysics: density field in material struct, int -> double
--This line, and those below, will be ignored--


SVN revision: 79851

+++ legacy/ephysics/src/lib/ephysics_body_materials.h
@@ -8,7 +8,7 @@ extern "C" {
 typedef struct _EPhysics_Body_Material_Props EPhysics_Body_Material_Props;
 struct _EPhysics_Body_Material_Props {
      EPhysics_Body_Material material;
-     int density;
+     double density;
      double friction;
      double restitution;
 };
